Add Basic IME Support for macOS
This PR added basic IME support for macOS. Particularly, a very simple
TextInputClient implementation added to facilitate the inputs.

Test Plan:

Run the example, add proper font support. Typing Chinese / Japanese in
the input box of the demo. See it now works.

TODO:

Need to position the IME window properly (likely to return reasonable
position in firstRectForCharacterRange:actualRange:).

Need to figure out whether need to add markedText support and how.
It requires to display the markedText, but not sure with imgui, what's
the best way of doing that (I want to avoid taint the input text, so
maybe it is another Text on screen somewhere? But how to do that not
intrusively?).